+++ public/index.mjs
@@ -1,4 +1,3 @@
-export { default as generate_eos_signature } from './generate_eos_signature.js'
+export { default as sign_txn } from './sign_txn.js'
 export { default as public_key_from_private } from './public_key_from_private.js'
 export { default as new_eos_keys } from './new_eos_keys.js'
-export { default as verify_eos_signature } from './verify_eos_signature.js'