+++ public/public_key_from_private.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const get_public_key = require('isomorphic-secp256k1/public/get_public_key')
+const { get_public_key } = require('universal-ecdsa')
 const public_key_to_wif = require('../private/public_key_to_wif')
 const wif_to_private_key = require('../private/wif_to_private_key')
 
@@ -26,13 +26,15 @@ const wif_to_private_key = require('../private/wif_to_private_key')
  * ```
  * @example <caption>Usage `public_key_from_private`.</caption>
  * ```js
- * public_key_from_private("5KQwrPbwdL6PhXujxW37FSSQZ1JiwsST4cqQzDeyXtP79zkvFD3").then(console.log)
+ * public_key_from_private(
+ *   '5KQwrPbwdL6PhXujxW37FSSQZ1JiwsST4cqQzDeyXtP79zkvFD3'
+ * ).then(console.log)
  * ```
  * The logged output will be EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV.
  */
 async function public_key_from_private(wif_private_key) {
   const private_key = await wif_to_private_key(wif_private_key)
-  return public_key_to_wif(await get_public_key({ private_key }))
+  return public_key_to_wif(await get_public_key(private_key))
 }
 
 module.exports = public_key_from_private