+++ changelog.md
@@ -6,9 +6,16 @@
 
 - Added `universal-ecdsa` package
 
+- Removed signature validation
+
+- Fix for invalid signature
+
 ### Minor
 
+- `new eos keys` function now contains a seed argument.
+
 - Removed `hard-coverage`
+
 - Removed `isomorphic-secp256k1`
 
 ### Patch